requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
discord.py==2.5.2
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
PyYAML==6.0.2
SQLAlchemy==2.0.41
typing_extensions==4.12.2
yarl==1.15.0
